Refactored code into packages
