--- src/data/duke.txt
@@ -1,4 +1,0 @@
-T | 1 | Task 1
-D | 1 | 2103T Submission | Aug 26 2022
-E | 0 | Meta Internship Breifing | Sep 2 2022
-E | 0 | ATAP Internship Breifing | Sep 1 2022
