+++ src/main/java/duke/Task.java
@@ -1,3 +1,5 @@
+package duke;
+
 public abstract class Task {
     protected String description;
     protected boolean isDone;
