+++ src/main/java/duke/command/Command.java
@@ -1,11 +1,17 @@
+package duke.command;
+
+import duke.DukeException;
+import duke.utils.Storage;
+import duke.TaskList;
+
 import java.io.IOException;
 
 public abstract class Command {
 
     /**
      * Runs the command based on the command type
-     * @param taskList TaskList to update tasks data
-     * @param storage Storage to save updates to TaskList
+     * @param taskList duke.TaskList to update tasks data
+     * @param storage duke.utils.Storage to save updates to duke.TaskList
      * @throws DukeException Index out of bounds / Improper syntax errors
      * @throws IOException Input Errors
      */
