+++ src/main/java/duke/command/ExitCommand.java
@@ -1,7 +1,13 @@
+package duke.command;
+
+import duke.utils.Storage;
+import duke.TaskList;
+import duke.utils.Ui;
+
 public class ExitCommand extends Command{
 
     /**
-     * Command to exit the application
+     * duke.command.Command to exit the application
      */
     @Override
     public void run(TaskList taskList, Storage storage) {
