+++ src/main/java/duke/command/PrintListCommand.java
@@ -1,9 +1,15 @@
+package duke.command;
+
+import duke.utils.Storage;
+import duke.TaskList;
+import duke.utils.Ui;
+
 public class PrintListCommand extends Command {
 
     /**
      * Prints all values of list currently
-     * @param taskList TaskList to print from
-     * @param storage Storage to store TaskList data to
+     * @param taskList duke.TaskList to print from
+     * @param storage duke.utils.Storage to store duke.TaskList data to
      */
     @Override
     public void run(TaskList taskList, Storage storage) {
