+++ src/main/java/duke/task/Todo.java
@@ -1,3 +1,7 @@
+package duke.task;
+
+import duke.Task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
